split test cases and make it more easy to maintain.
